--- client/.env.production
@@ -1,6 +1,0 @@
-NEXT_PUBLIC_CLIENT=https://rando.gg
-NEXT_PUBLIC_SERVER=https://rando.gg
-NEXT_PUBLIC_TURN=turn:165.232.150.149:3478
-NEXT_PUBLIC_TURN_FORCE_RELAY=false
-NEXT_PUBLIC_TURN_SERVER_USER=test
-NEXT_PUBLIC_TURN_SERVER_PASS=test123
