+++ client/env.d.ts
@@ -3,11 +3,10 @@ declare global {
     interface ProcessEnv {
       NODE_ENV: 'development' | 'production'
       APP_ENV: 'test' | undefined
-      NEXT_PUBLIC_CLIENT: string
-      NEXT_PUBLIC_SERVER: string
-      NEXT_PUBLIC_TURN: string
-      NEXT_PUBLIC_TURN_SERVER_USER: string
-      NEXT_PUBLIC_TURN_SERVER_PASS: string
+      SIGNALR_SERVER_URL: string
+      TURN_SERVER: string
+      TURN_SERVER_SERVER_USER: string
+      TURN_SERVER_SERVER_PASS: string
     }
   }
 }
