+++ client/webpack.common.config.ts
@@ -1,0 +1,50 @@
+import path from 'path'
+import { Configuration, HotModuleReplacementPlugin } from 'webpack'
+import Dotenv from 'dotenv-webpack'
+import HtmlWebpackPlugin from 'html-webpack-plugin'
+
+const config: Configuration = {
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    publicPath: '/',
+    filename: 'bundle.js',
+    assetModuleFilename: 'src/assets/images/[name].[ext]',
+    clean: true
+  },
+  entry: '/src/index.tsx',
+  module: {
+    rules: [
+      {
+        test: /\.(ts|js)x?$/i,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env', '@babel/preset-react', '@babel/preset-typescript'],
+          },
+        },
+      },
+      {
+        test: /\.css$/i,
+        use: ['style-loader', 'css-loader'],
+      },
+      {
+        test: /\.(png|jpg|gif|svg|eot|ttf|woff)$/,
+        type: 'asset/resource'
+     }
+    ],
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js'],
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: 'src/index.html',
+      favicon: 'src/assets/favicon.ico'
+    }),
+    new HotModuleReplacementPlugin(),
+    new Dotenv(),
+  ],
+}
+
+export default config
