+++ client/webpack.dev.config.ts
@@ -1,65 +1,25 @@
 import path from 'path'
-import { Configuration as WebpackConfiguration, HotModuleReplacementPlugin } from 'webpack'
-import Dotenv from 'dotenv-webpack'
+import { Configuration as WebpackConfiguration } from 'webpack'
 import { Configuration as WebpackDevServerConfiguration } from 'webpack-dev-server'
-import HtmlWebpackPlugin from 'html-webpack-plugin'
+import { merge } from 'webpack-merge'
+import common from './webpack.common.config'
 
 interface Configuration extends WebpackConfiguration {
   devServer?: WebpackDevServerConfiguration
 }
 
-const config: Configuration = {
+const config: Configuration = merge(common, {
   mode: 'development',
-  output: {
-    path: '/dist',
-    filename: 'bundle.js',
-    publicPath: '/',
-    assetModuleFilename: 'src/assets/images/[name].[ext]'
-  },
-  entry: '/src/index.tsx',
-  module: {
-    rules: [
-      {
-        test: /\.(ts|js)x?$/i,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: ['@babel/preset-env', '@babel/preset-react', '@babel/preset-typescript'],
-          },
-        },
-      },
-      {
-        test: /\.css$/i,
-        use: ['style-loader', 'css-loader'],
-      },
-      {
-        test: /\.(png|jpg|gif|svg|eot|ttf|woff)$/,
-        type: 'asset/resource'
-     }
-    ],
-  },
-  resolve: {
-    extensions: ['.tsx', '.ts', '.js'],
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: 'src/index.html',
-      favicon: 'src/assets/favicon.ico'
-    }),
-    new HotModuleReplacementPlugin(),
-    new Dotenv(),
-  ],
   devtool: 'inline-source-map',
   devServer: {
-    static: path.join(__dirname, 'dist'),
+    static: path.resolve(__dirname, 'dist'),
     port: 3000,
     historyApiFallback: {
-      index: '/404'
+      index: '/404',
     },
     open: false,
     hot: false,
   },
-}
+})
 
 export default config
