--- docker-compose.yml
@@ -1,18 +1,0 @@
-version: '3.8'
-services:
-  client:
-    build: ./client
-    image: vandervillain/rando:client
-    depends_on:
-      - signalr
-    container_name: client
-    ports:
-      - '3000:3000'
-  signalr:
-    build: ./azure-function
-    image: vandervillain/rando:signalr
-    container_name: signalr
-    ports:
-      - '5000:5000'
-    env_file:
-      - azure-function/env.list
